If a wake key is disabled by the user, do not wake from doze.
Currently, any wake key will wake the device from a doze, even if that
key has not been enabled as a wake key in Settings.

If the device is 'dreaming' in the Doze state, check if the user has
explicitly disabled the wake key (or never enabled the setting in the
first place) before waking the device.

Adapt for Mark

Change-Id: I7397087c143161e8e1ddb84d0e23f6027fea0aac